Move location props and mapping to LocationLists component
