+++ desktop/packages/mullvad-vpn/src/renderer/components/views/select-location/components/recent-locations/components/recent-location-list/hooks/use-recent-locations.ts
@@ -1,14 +1,12 @@
 import { LocationType } from '../../../../../../../../features/locations/types';
-import { useSelectLocationViewContext } from '../../../../../SelectLocationViewContext';
+import { useLocationListsContext } from '../../../../location-lists/LocationListsContext';
 
 export function useRecentLocations() {
-  const { locationType, recentEntryLocations, recentExitLocations } =
-    useSelectLocationViewContext();
-  if (recentEntryLocations && locationType === LocationType.entry) {
+  const { type, recentEntryLocations, recentExitLocations } = useLocationListsContext();
+  if (recentEntryLocations && type === LocationType.entry) {
     return recentEntryLocations;
-  } else if (recentExitLocations && locationType === LocationType.exit) {
+  } else if (recentExitLocations && type === LocationType.exit) {
     return recentExitLocations;
   }
-
   return [];
 }
